\name{wagepan}
\alias{wagepan}
\docType{data}
\title{
Data to illustrate mixed effects regression models with serial correlation.
}
\description{
Replication data illustrating serial correlation specifications to adjust for correlated residuals.
}
\usage{data("wagepan")}
\format{
  A data frame with 4360 observations on the following 51 variables.
  \describe{
    \item{\code{nr}}{a numeric vector}
    \item{\code{year}}{a numeric vector}
    \item{\code{agric}}{a numeric vector}
    \item{\code{black}}{a numeric vector}
    \item{\code{bus}}{a numeric vector}
    \item{\code{construc}}{a numeric vector}
    \item{\code{ent}}{a numeric vector}
    \item{\code{exper}}{a numeric vector}
    \item{\code{fin}}{a numeric vector}
    \item{\code{hisp}}{a numeric vector}
    \item{\code{poorhlth}}{a numeric vector}
    \item{\code{hours}}{a numeric vector}
    \item{\code{manuf}}{a numeric vector}
    \item{\code{married}}{a numeric vector}
    \item{\code{min}}{a numeric vector}
    \item{\code{nrthcen}}{a numeric vector}
    \item{\code{nrtheast}}{a numeric vector}
    \item{\code{occ1}}{a numeric vector}
    \item{\code{occ2}}{a numeric vector}
    \item{\code{occ3}}{a numeric vector}
    \item{\code{occ4}}{a numeric vector}
    \item{\code{occ5}}{a numeric vector}
    \item{\code{occ6}}{a numeric vector}
    \item{\code{occ7}}{a numeric vector}
    \item{\code{occ8}}{a numeric vector}
    \item{\code{occ9}}{a numeric vector}
    \item{\code{per}}{a numeric vector}
    \item{\code{pro}}{a numeric vector}
    \item{\code{pub}}{a numeric vector}
    \item{\code{rur}}{a numeric vector}
    \item{\code{south}}{a numeric vector}
    \item{\code{educ}}{a numeric vector}
    \item{\code{tra}}{a numeric vector}
    \item{\code{trad}}{a numeric vector}
    \item{\code{union}}{a numeric vector}
    \item{\code{lwage}}{a numeric vector}
    \item{\code{d81}}{a numeric vector}
    \item{\code{d82}}{a numeric vector}
    \item{\code{d83}}{a numeric vector}
    \item{\code{d84}}{a numeric vector}
    \item{\code{d85}}{a numeric vector}
    \item{\code{d86}}{a numeric vector}
    \item{\code{d87}}{a numeric vector}
    \item{\code{expersq}}{a numeric vector}
    \item{\code{r}}{a numeric vector}
    \item{\code{num}}{a numeric vector}
    \item{\code{number}}{a numeric vector}
    \item{\code{mn_lwage}}{a numeric vector}
    \item{\code{yeart}}{a numeric vector}
    \item{\code{yeart2}}{a numeric vector}
    \item{\code{yeart3}}{a numeric vector}
  }
}
\examples{
data(wagepan)
head(wagepan)
}
\keyword{datasets}
